# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a retail management system server.",
        "title": "Retail Pro API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/categories": {
            "get": {
                "description": "Retrieves a list of all categories associated with the logged-in user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Get all categories for the authenticated user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/routes_category.Category"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: User ID not found in context",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new category for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Create a new category",
                "parameters": [
                    {
                        "description": "Category details",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes_category.CreateCategory"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully retrieved categories",
                        "schema": {
                            "$ref": "#/definitions/routes_category.Category"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: User ID not found in context",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "409": {
                        "description": "Category with this name already exists",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    }
                }
            }
        },
        "/categories/name/{name}": {
            "get": {
                "description": "Retrieves a specific category by its name for the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Get category by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved categories",
                        "schema": {
                            "$ref": "#/definitions/routes_category.Category"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: User ID not found in context",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    }
                }
            }
        },
        "/categories/{id}": {
            "get": {
                "description": "Retrieves a specific category by its ID for the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Get category by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved categories",
                        "schema": {
                            "$ref": "#/definitions/routes_category.Category"
                        }
                    },
                    "400": {
                        "description": "Invalid category ID format (if applicable)",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: User ID not found in context",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing category by its ID for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Update an existing category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated category details",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes_category.UpdateCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes_category.Category"
                        }
                    },
                    "400": {
                        "description": "Invalid request data or ID format",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: User ID not found in context or not owner",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "409": {
                        "description": "Category with this name already exists",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a category by its ID for the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Delete a category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes_category.Category"
                        }
                    },
                    "400": {
                        "description": "Invalid category ID format",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: User ID not found in context or not owner",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    }
                }
            }
        },
...
        "/products/category/{categoryID}": {
            "get": {
                "description": "Retrieves all products belonging to a specific Category ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get products by Category ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "categoryID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved products",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/routes_product.Product"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Category ID format",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_yantology_simple-pos_pkg_dto.MessageResponse": {
            "description": "Generic message response model",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                }
            }
        },
...
        "routes_category.Category": {
            "description": "Category model",
            "type": "object",
            "required": [
                "created_at",
                "id",
                "name",
                "updated_at",
                "user_id"
            ],
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-04-25T15:04:05Z07:00"
                },
                "id": {
                    "description": "Changed from string to int",
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Electronics"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-04-25T15:04:05Z07:00"
                },
                "user_id": {
                    "description": "Changed from string to int",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "routes_category.CreateCategory": {
            "description": "Create category request model",
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Groceries"
                }
            }
        },
        "routes_category.UpdateCategoryRequest": {
            "description": "Update category request model",
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Home Goods"
                }
            }
        },
...
        "routes_product.Product": {
            "description": "Product model",
            "type": "object",
            "properties": {
                "category_id": {
                    "description": "Changed from string to int",
                    "type": "integer",
                    "example": 1
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-04-25T15:04:05Z07:00"
                },
                "id": {
                    "description": "Changed from string to int",
                    "type": "integer",
                    "example": 1
                },
                "is_available": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "Laptop Pro"
                },
                "price": {
                    "type": "number",
                    "example": 15000000
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-04-25T15:04:05Z07:00"
                },
                "user_id": {
                    "description": "Changed from string to int",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}